api/requirements.txt
flask
requests
flask-cors
redis
pyopenssl
flask_mongoengine
pyjwt
bcrypt
